Improve OInK build times
