+++ roboplan_oink/include/roboplan_oink/optimal_ik.hpp
@@ -10,37 +10,13 @@
 #include <roboplan/core/scene.hpp>
 #include <roboplan/core/types.hpp>
 
-// Forward declaration of the ProxQP dense solver. The solver is a private implementation
-// detail of roboplan_oink; consumers of this header never need the proxsuite headers.
-namespace proxsuite::proxqp::dense {
-template <typename T> struct QP;
-}  // namespace proxsuite::proxqp::dense
+#include <roboplan_oink/detail/qp_solver_fwd.hpp>
+#include <roboplan_oink/oink_settings.hpp>
 
 namespace roboplan {
 
 /// @brief Infinity value used for unbounded QP constraint bounds.
 constexpr double kInfinity = std::numeric_limits<double>::infinity();
-
-/// @brief Solver settings for the Oink QP solver (ProxQP).
-///
-/// This is a solver-agnostic subset of the underlying ProxQP settings. See
-/// https://simple-robotics.github.io/proxsuite/ for the semantics of the tolerances.
-struct OinkSettings {
-  /// Absolute stopping tolerance on the primal/dual residuals.
-  double eps_abs = 1e-6;
-  /// Relative stopping tolerance on the primal/dual residuals (0 disables it).
-  double eps_rel = 0.0;
-  /// Maximum number of solver iterations.
-  int max_iter = 10000;
-  /// Print solver internals to stdout.
-  bool verbose = false;
-  /// Warm start each solve with the previous solution (recommended for control loops).
-  bool warm_start = true;
-  /// When the QP is primal-infeasible (e.g., a violated barrier conflicting with velocity
-  /// limits), solve the closest feasible problem in the least-squares sense instead of
-  /// failing. This guarantees that solveIk() always returns a usable displacement.
-  bool primal_infeasibility_solving = true;
-};
 
 /// @brief Abstract base class for IK tasks.
 ///
@@ -309,7 +285,6 @@ struct Oink {
   Oink(const Scene& scene, const OinkSettings& custom_settings)
       : Oink(scene, "", custom_settings) {}
 
-  /// Out-of-line destructor required because the ProxQP solver is only forward-declared here.
   ~Oink();
 
   /// @brief Solve inverse kinematics for tasks only
@@ -446,7 +421,7 @@ private:
 public:
   // QP solver (ProxQP dense backend). Reconstructed whenever the constraint row count
   // changes, since ProxQP fixes the problem dimensions at construction.
-  std::unique_ptr<proxsuite::proxqp::dense::QP<double>> solver;
+  detail::QpSolverPtr solver;
   OinkSettings settings;
 
   // Problem dimensions
