+++ roboplan_oink/src/optimal_ik.cpp
@@ -2,8 +2,8 @@
 #include <limits>
 
 #include <pinocchio/algorithm/joint-configuration.hpp>
-#include <proxsuite/proxqp/dense/dense.hpp>
 #include <roboplan_oink/optimal_ik.hpp>
+#include <roboplan_oink/qp_backend.hpp>
 
 namespace {
 // Minimum squared norm threshold to avoid division by zero in barrier regularization
@@ -169,8 +169,6 @@ Oink::Oink(const Scene& scene, const std::string& group_name)
   collision_context_ = std::make_unique<CollisionContext>(scene);
 }
 
-// Defined here (not in the header) because proxsuite::proxqp::dense::QP is only
-// forward-declared in the header.
 Oink::~Oink() = default;
 
 tl::expected<void, std::string>
@@ -336,75 +334,9 @@ Oink::solveIk(const Scene& scene, const std::vector<std::shared_ptr<Task>>& task
   constraint_sizes.clear();
   barrier_sizes.clear();
 
-  using proxsuite::nullopt;
-  using proxsuite::proxqp::InitialGuessStatus;
-  using proxsuite::proxqp::QPSolverOutput;
-
-  if (init_required) {
-    // ProxQP fixes the problem dimensions at construction, so the solver is rebuilt
-    // whenever the number of constraint rows changes.
-    solver = std::make_unique<proxsuite::proxqp::dense::QP<double>>(num_variables, /*n_eq=*/0,
-                                                                    /*n_in=*/total_rows);
-    solver->settings.eps_abs = settings.eps_abs;
-    solver->settings.eps_rel = settings.eps_rel;
-    solver->settings.max_iter = settings.max_iter;
-    solver->settings.verbose = settings.verbose;
-    solver->settings.primal_infeasibility_solving = settings.primal_infeasibility_solving;
-    solver->settings.initial_guess = InitialGuessStatus::NO_INITIAL_GUESS;
-
-    if (total_rows > 0) {
-      solver->init(H, c, nullopt, nullopt, constraint_workspace_A, constraint_workspace_lower,
-                   constraint_workspace_upper);
-    } else {
-      solver->init(H, c, nullopt, nullopt, nullopt, nullopt, nullopt);
-    }
-  } else {
-    if (total_rows > 0) {
-      solver->update(H, c, nullopt, nullopt, constraint_workspace_A, constraint_workspace_lower,
-                     constraint_workspace_upper);
-    } else {
-      solver->update(H, c, nullopt, nullopt, nullopt, nullopt, nullopt);
-    }
-  }
-
-  solver->solve();
-
-  // PROXQP_SOLVED_CLOSEST_PRIMAL_FEASIBLE is returned when the QP was primal-infeasible and
-  // settings.primal_infeasibility_solving is enabled: the solution minimizes the constraint
-  // violation in the least-squares sense and is the safest displacement available.
-  const QPSolverOutput status = solver->results.info.status;
-  if (status != QPSolverOutput::PROXQP_SOLVED &&
-      status != QPSolverOutput::PROXQP_SOLVED_CLOSEST_PRIMAL_FEASIBLE) {
-    // The solve did not converge, so results.x is not a trustworthy displacement. Reset the
-    // initial guess so a subsequent solveIk() does not warm start from this bad solution (the
-    // equivalent of OsqpEigen::Solver::clearSolverVariables()).
-    solver->settings.initial_guess = InitialGuessStatus::NO_INITIAL_GUESS;
-    switch (status) {
-    case QPSolverOutput::PROXQP_MAX_ITER_REACHED:
-      return tl::make_unexpected("QP solver reached the maximum number of iterations");
-    case QPSolverOutput::PROXQP_PRIMAL_INFEASIBLE:
-      return tl::make_unexpected(
-          "QP is primal infeasible (constraints and barriers cannot be satisfied "
-          "simultaneously). Enable OinkSettings::primal_infeasibility_solving to compute the "
-          "closest feasible solution instead.");
-    case QPSolverOutput::PROXQP_DUAL_INFEASIBLE:
-      return tl::make_unexpected("QP is dual infeasible (objective unbounded below)");
-    default:
-      return tl::make_unexpected("QP solver failed to find a solution");
-    }
-  }
-
-  // Warm start subsequent solves with the previous solution. This must only be enabled
-  // after a successful solve has populated the results and factorization; enabling it before
-  // the first solve (or after a failed one) makes ProxQP reuse a factorization/solution that
-  // was never validly computed.
-  if (settings.warm_start) {
-    solver->settings.initial_guess = InitialGuessStatus::WARM_START_WITH_PREVIOUS_RESULT;
-  }
-
-  // Extract the solution and copy into delta_q
-  delta_q.noalias() = solver->results.x;
-  return {};
+  return detail::solveQp(solver, settings, init_required, num_variables, total_rows, H, c,
+                         constraint_workspace_A, constraint_workspace_lower,
+                         constraint_workspace_upper, delta_q);
 }
 
 // Overload: tasks only
@@ -444,8 +376,11 @@ Oink::enforceBarriers(const Scene& scene, const std::vector<std::shared_ptr<Barr
   const auto& model = scene.getModel();
   const Eigen::VectorXd& q = scene.getCurrentJointPositions();
 
-  // Compute candidate configuration by integrating delta_q
-  const Eigen::VectorXd q_candidate = pinocchio::integrate(model, q, delta_q);
+  // Compute candidate configuration by integrating delta_q. The copy into a plain VectorXd
+  // makes the call match pinocchio's pre-instantiated integrate() signature, so this TU does
+  // not instantiate the joint-visitor templates itself.
+  const Eigen::VectorXd dq(delta_q);
+  const Eigen::VectorXd q_candidate = pinocchio::integrate(model, q, dq);
 
   // Evaluate all barriers at the candidate configuration. enforce_barriers_data is a
   // pre-allocated pinocchio::Data scoped to this method, so we don't mutate scene state.
